updates to new service pre-tested
